// repos/tvl-labs/swap-sdk/balancer-js/src/types.ts
export interface ContractAddresses {
  vault: string;
  multicall: string;
  lidoRelayer?: string;
  gaugeController?: string;
  feeDistributor?: string;
  protocolFeePercentagesProvider?: string;
}

export interface BalancerNetworkConfig {
  chainId: Network;
  addresses: {
    contracts: ContractAddresses;
    tokens: {
      wrappedNativeAsset: string;
      lbpRaisingTokens?: string[];
      stETH?: string;
      wstETH?: string;
      bal?: string;
      veBal?: string;
      bbaUsd?: string;
    };
  };
  urls: {
    subgraph: string;
    gaugesSubgraph?: string;
    blockNumberSubgraph?: string;
  };
  thirdParty: {
    coingecko: {
      nativeAssetId: string;
      platformId: string;
    };
  };
  averageBlockTime?: number; // In seconds, used if blockNumberSubgraph not set
  pools: {
    wETHwstETH?: PoolReference;
  };
  poolsToIgnore?: string[];
  sorConnectingTokens?: { symbol: string; address: string }[];
  sorTriPathMidPoolIds?: string[];
}
// ...
export type PoolReference = {
  id: string;
  address: string;
};